Rename project to wagtail_polymath
