+++ MANIFEST.in
@@ -1,0 +1,6 @@
+include CONTRIBUTORS.md
+include CONTRIBUTING.md
+include CHANGELOG.md
+include LICENSE
+include README.md
+recursive-include wagtail_polymath *.html *.png *.gif *js *.css *jpg *jpeg *svg *py
