+++ docs/index.rst
@@ -3,7 +3,7 @@
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Welcome to wagtailmath's documentation!
+Welcome to wagtail-polymath's documentation!
 =================================================================
 
 Contents:
