+++ src/wagtail_polymath/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class WagtailPolymathConfig(AppConfig):
+    name = "wagtail_polymath"
