+++ src/wagtail_polymath/static/wagtail_polymath/js/wagtailmath.js
