+++ src/wagtail_polymath/templates/wagtail_polymath/mathjaxwidget.html
