+++ tests/testapp/templates/testapp/math_page.html
@@ -1,4 +1,4 @@
-{% load wagtailmath %}
+{% load wagtail_polymath %}
 <html>
 <head>
     <script src="{% mathjax %}"></script>
